+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from 'react';
 
 import { Providers, ThemeProvider } from '@/components/layout';
 import type { Metadata } from 'next';
-import { Geist_Mono, Lato } from 'next/font/google';
+import { Geist_Mono, Lato, Space_Grotesk } from 'next/font/google';
 import './globals.css';
 
 const lato = Lato({
@@ -14,6 +14,12 @@ const lato = Lato({
 const geistMono = Geist_Mono({
   variable: '--font-geist-mono',
   subsets: ['latin'],
+});
+
+const spaceGrotesk = Space_Grotesk({
+  variable: '--font-space-grotesk',
+  subsets: ['latin'],
+  weight: ['500', '600', '700'],
 });
 
 export const metadata: Metadata = {
@@ -28,7 +34,9 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en" suppressHydrationWarning data-scroll-behavior="smooth" className="no-scrollbar">
-      <body className={`${lato.variable} ${geistMono.variable} relative antialiased`}>
+      <body
+        className={`${lato.variable} ${geistMono.variable} ${spaceGrotesk.variable} relative antialiased`}
+      >
         <ThemeProvider attribute="class" defaultTheme="light" disableTransitionOnChange>
           <Providers>{children}</Providers>
         </ThemeProvider>